@@ -17,7 +17,7 @@
   "devDependencies": {
     "@babel/cli": "^7.15.7",
     "@babel/core": "^7.15.7",
-    "@vitejs/plugin-react": "^1.0.2",
+    "@vitejs/plugin-react": "^1.3.1",
     "babel-eslint": "^10.1.0",
     "babel-loader": "^8.2.2",
     "concurrently": "^6.2.1",
@@ -38,10 +38,15 @@
     "vite": "^2.5.10"
   },
   "dependencies": {
-    "dotenv": "^10.0.0",
+    "@chakra-ui/react": "^2.0.0",
+    "@emotion/react": "^11.9.0",
+    "@emotion/styled": "^11.8.1",
+    "dotenv": "^16.0.0",
     "express": "^4.17.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.3.0"
+    "framer-motion": "^6.3.3",
+    "knex": "^1.0.7",
+    "react": "^18.1.0",
+    "react-dom": "^18.0.0",
+    "react-router-dom": "^6.3.0"
   }
 }
